# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API маркетплейса Клевер (микросервисная архитектура).\n- Auth Service       (HTTP :8001, gRPC :9001) — авторизация, профиль, пользователи\n- Support Service    (HTTP :8002)             — техподдержка\n- Catalog Service    (HTTP :8004, gRPC :9004) — объявления, категории, избранное, просмотры\n- Commerce Service   (HTTP :8006)             — корзина, чаты, заказы\nВсе запросы идут через API Gateway (nginx :8080).",
        "title": "Clover API",
        "contact": {},
        "version": "1.0"
    },
    "host": "clover-go.ru",
    "basePath": "/api/v1",
    "paths": {
...
        "/ads/{id}/promotions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotion"
                ],
                "summary": "Активные промо объявления",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID объявления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PromotionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid ad id",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotion"
                ],
                "summary": "Купить продвижение объявления",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID объявления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "plan_code и idempotency_key",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PurchasePromotionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PurchasePromotionResponse"
                        }
                    },
                    "400": {
                        "description": "INSUFFICIENT_FUNDS / PLAN_NOT_FOUND / PLAN_INACTIVE / INVALID_AD_STATUS / INVALID_REQUEST",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "NOT_AD_OWNER",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "AD_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ads/{id}/view": {
            "post": {
                "description": "Записывает просмотр с дедупликацией и возвращает актуальный счётчик",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ads"
                ],
                "summary": "Зафиксировать просмотр объявления",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID объявления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID v4 идентификатор устройства",
                        "name": "X-Device-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "views_count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid ad id / invalid X-Device-ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "ad not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/profile/promotions": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotion"
                ],
                "summary": "История покупок продвижения",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "лимит (по умолчанию 20, максимум 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID последнего элемента предыдущей страницы",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PromotionListResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/promotion/plans": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promotion"
                ],
                "summary": "Список тарифов продвижения",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PromotionPlanResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "dto.PromotionListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PromotionResponse"
                    }
                },
                "next_cursor": {
                    "type": "integer"
                }
            }
        },
        "dto.PromotionPlanResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "duration_days": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "dto.PromotionResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "plan_code": {
                    "type": "string"
                },
                "price_paid": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "starts_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.PurchasePromotionRequest": {
            "type": "object",
            "properties": {
                "idempotency_key": {
                    "type": "string"
                },
                "plan_code": {
                    "type": "string"
                }
            }
        },
        "dto.PurchasePromotionResponse": {
            "type": "object",
            "properties": {
                "promotion": {
                    "$ref": "#/definitions/dto.PromotionResponse"
                },
                "wallet_balance": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "JWT session token",
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        },
        "CsrfCookieAuth": {
            "description": "CSRF token stored in cookies",
            "type": "apiKey",
            "name": "csrf_token",
            "in": "cookie"
        },
        "CsrfHeaderAuth": {
            "description": "CSRF token required in header",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}